backend/pyproject.toml
[tool.poetry]
name = "django-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
django-rest-framework = "^0.1.0"
strawberry-graphql = "^0.227.2"
strawberry-graphql-django = "^0.39.2"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
